+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - Cache installationID async to avoid file IO on the main thread when starting the SDK (#3601)
 
+### Fixes
+
+- Don't use main thread for app hang screenshot and view hierarchy (#3547)
 
 ## 8.20.0